move endianess enum
